+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -412,6 +412,7 @@ sub build_isf_public_interface_contract {
             't/1246-isf-setter-syntax.t',
             't/1247-isf-clock-domain-partition.t',
             't/1248-isf-rule-trigger-parameter-binding.t',
+            't/1249-isf-activation-parameter-constants.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',